abstract out dark mode toggle switch
