+++ src/navBar/NavBar.js
@@ -1,21 +1,17 @@
-import React, { useMemo, useContext } from 'react';
+import React, { useContext } from 'react';
 
-import Switch from 'react-switch';
 import Popup from 'reactjs-popup';
 import CtaButton from '../components/buttons/CtaButton';
-import { TbSun, TbMoon } from 'react-icons/tb';
 
 import ThemeContext from '../context/ThemeContext';
 import ModalContainer from '../components/containers/ModalContainer';
 import classNames from 'classnames';
+import DarkModeToggle from './components/DarkModeToggle';
 
 const NavBar = ({ handleChange, theme }) => {
   const overlayStyle = { background: 'rgba(0,0,0,0.5)' };
 
   const themeColor = useContext(ThemeContext);
-  const handleColor = useMemo(() => {
-    return themeColor === 'dark' ? '#1D1E1E' : '#E9EDF0';
-  }, [themeColor]);
 
   const handleScroll = () => {
     const element = document.getElementById('Projects');
@@ -30,24 +26,7 @@ const NavBar = ({ handleChange, theme }) => {
           'p-1 h-8 rounded-lg'
         )}
       >
-        <Switch
-          borderRadius={5}
-          height={22}
-          width={50}
-          boxShadow={false}
-          onChange={handleChange}
-          checked={theme === 'dark'}
-          offHandleColor={handleColor}
-          onHandleColor={handleColor}
-          offColor="#2EB2EC"
-          onColor="#2EB2EC"
-          uncheckedIcon={
-            <TbSun className="mx-auto text-neoBackground translate-y-[3px]" />
-          }
-          checkedIcon={
-            <TbMoon className="mx-auto text-neoBackground translate-y-[3px]" />
-          }
-        />
+        <DarkModeToggle handleChange={handleChange} />
       </div>
 
       <div className="flex flex-row space-x-8 items-center">
